requirements.txt
pandas
scikit-learn
transformers
datasets
# torch
accelerate>=0.26.0
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
python-dotenv
openai
cachetools
faiss-cpu
sentence-transformers
# lxml_html_clean
# newspaper3k
sqlalchemy
fastapi
uvicorn
psycopg2-binary